#95 - Update Jest configuration, refactor test imports, and add Babel configuration
