+++ webapp/babel.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    presets: [
+      '@babel/preset-env',
+      '@babel/preset-react'
+    ],
+  };
