+++ webapp/e2e/jest.config.js
@@ -1,4 +1,7 @@
 module.exports = {
+    transform: {
+        '^.+\\.jsx?$': 'babel-jest',
+    },
     testMatch: ["**/steps/*.js"],
     testTimeout: 30000,
     setupFilesAfterEnv: ["expect-puppeteer"]
