+++ webapp/src/tests/InGameChat.test.js
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom"
 import { render, screen, fireEvent, waitFor } from "@testing-library/react"
-import InGameChat from "./InGameChat"
+import InGameChat from "../app/game/InGameChat"
 import fetchMock from "jest-fetch-mock"
 import { act } from "react"
